clean up configuration requests
    fix plugin loading to avoid loading plugins twice in certain circumstances
    use new map binder based input registration, allow for Graylog2Module being a superclass for reflection-based loading of guice modules

    fixes Graylog2/graylog2-web-interface#944

+++ dataset/GitHub_Java/Graylog2.graylog2-server/graylog2-inputs/src/main/java/org/graylog2/inputs/transports/AmqpTransport.java
@@ -137,11 +137,11 @@ public class AmqpTransport extends ThrottleableTransport {
     public void launch(MessageInput input) throws MisfireException {
         consumer = new AmqpConsumer(
                 configuration.getString(CK_HOSTNAME),
-                configuration.getInt(CK_PORT),
+                (int) configuration.getInt(CK_PORT),
                 configuration.getString(CK_VHOST),
                 configuration.getString(CK_USERNAME),
                 configuration.getString(CK_PASSWORD),
-                configuration.getInt(CK_PREFETCH),
+                (int) configuration.getInt(CK_PREFETCH),
                 configuration.getString(CK_QUEUE),
                 configuration.getString(CK_EXCHANGE),
                 configuration.getString(CK_ROUTING_KEY),
